Refactor to read emails as well as names from file
Specifically, the read_names() function now uses a helper function
read_line() to tokenize each line of the file and assign an email to
the STUDENT_DATA struct only if it is found.

Add pre-release compiler directive, currently used to select data file
with emails instead of the names file.